requirements.txt
flask
flask_cors
rasa
pymongo
python-dotenv
beautifulsoup4
pycryptodome
google
sanic
PyJWT
spacy
sklearn
pandas
numpy
psutil
flask-pymongo
pillow
bcrypt
pyfladesk
flask-executor
ipfshttpclient
flask-socketio
